Fixed bugs with settings and import
